requirements.txt
pyyaml==3.13
paho-mqtt==1.4.0
setproctitle==1.1.10